+++ CHANGELOG.md
@@ -325,3 +325,5 @@
 * Define `top` expression
 * Define `sub` expression
 * Define `rep` expression
+* Define `Forest` type family
+* Define `Twice` type alias